// client/stocklinx app/src/redux/user/constant.ts
export enum userConst {
  FETCH_USERS_REQUEST = "FETCH_USERS_REQUEST",
  FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS",
  FETCH_USERS_FAILURE = "FETCH_USERS_FAILURE",
  FETCH_USER_REQUEST = "FETCH_USER_REQUEST",
  FETCH_USER_SUCCESS = "FETCH_USER_SUCCESS",
  FETCH_USER_FAILURE = "FETCH_USER_FAILURE",
  CREATE_USER_REQUEST = "CREATE_USER_REQUEST",
  CREATE_USER_SUCCESS = "CREATE_USER_SUCCESS",
  CREATE_USER_FAILURE = "CREATE_USER_FAILURE",
  UPDATE_USER_REQUEST = "UPDATE_USER_REQUEST",
  UPDATE_USER_SUCCESS = "UPDATE_USER_SUCCESS",
  UPDATE_USER_FAILURE = "UPDATE_USER_FAILURE",
  REMOVE_USER_REQUEST = "REMOVE_USER_REQUEST",
  REMOVE_USER_SUCCESS = "REMOVE_USER_SUCCESS",
  REMOVE_USER_FAILURE = "REMOVE_USER_FAILURE",
  SIGN_IN_REQUEST = "SIGN_IN_REQUEST",
  SIGN_IN_SUCCESS = "SIGN_IN_SUCCESS",
  SIGN_IN_FAILURE = "SIGN_IN_FAILURE",
  GET_WITH_TOKEN_REQUEST = "GET_WITH_TOKEN_REQUEST",
  GET_WITH_TOKEN_SUCCESS = "GET_WITH_TOKEN_SUCCESS",
  GET_WITH_TOKEN_FAILURE = "GET_WITH_TOKEN_FAILURE",
}